fix(ci): enforce pytest exit status in test-python; make fast suite pass torch-less
The test-python job was false-green: `pytest ... | tee` without pipefail
made the step's exit status tee's (always 0), so collection aborts and
test failures were silently swallowed — zero Python tests gated CI.

Workflow:
- Add `set -o pipefail` to the "Run fast tests" and "Run integration
  tests" steps (both pipe through tee).
- Lower --cov-fail-under from 75 to 45 with rationale: the dev-only job
  has no torch, so all RL-gated modules measure 0%; real coverage in
  this environment is 48% and the 75 gate was never actually enforced.

Collection (18 errors torch-less):
- Add module-level `pytest.importorskip("torch")` guards (existing
  tests/test_training.py convention) to 17 modules that import torch
  directly or transitively.
- tests/test_code_generation.py: importorskip the *generated* module
  (bucket_brigade.agents.archetypes_generated is produced by
  scripts/generate_python.py and is not committed).

Newly-exposed failures (diagnosed, not just widened):
- tests/test_evolution.py::test_ga_evolve_basic: STALE EXPECTATION.
  Fitness is GameResult.final_score; since bucket-brigade-core 354219d9
  ("Align Rust reward computation with Python per-step logic",
  deliberate BREAKING change) all reward components accrue per step, so
  final_score scales with episode length x num_agents (~5218 for
  trivial_cooperation, not the pre-354219d9 terminal-payoff range
  [-100, 100]). Bound is now derived from the scenario parameters.
- tests/test_evolution.py::test_parallel_with_scenario: STALE API.
  RustFitnessEvaluator dropped the `scenario=` object kwarg for
  `scenario_name` + `num_agents` in 0e2fdb30.
- bucket_brigade/equilibrium/payoff_evaluator.py: sequential branch
  returned np.float32 (float32 reward arrays promote the accumulator),
  violating its declared `-> float` contract; parallel branch already
  cast. One-line float() cast.
- tests/test_kstar_vs_trainability.py: byte-identical JSON regeneration
  does not survive numpy 1.26 -> 2.x reduction changes (last-ulp float
  noise only; the committed .md still matches byte-for-byte). Compare
  the JSON numerically with rel=1e-9; structure/ints/strings exact.

Torch-less fast suite: 836 passed, 29 skipped, 0 errors, coverage 48%.

Closes #484

+++ bucket_brigade/equilibrium/payoff_evaluator.py
@@ -176,7 +176,12 @@ class PayoffEvaluator:
 
                 total_reward += episode_reward
 
-            return total_reward / self.num_simulations
+            # Cast to a builtin float: `rewards` is a float32 numpy array, so
+            # the `episode_reward`/`total_reward` accumulation above promotes
+            # to np.float32 (which is NOT an instance of `float`, unlike
+            # np.float64). The declared return type is `float` and the
+            # parallel branch already casts (issue #484).
+            return float(total_reward / self.num_simulations)
 
     def evaluate_payoff_matrix(
         self,
